+++ NEWS.md
@@ -1,5 +1,15 @@
 
 
+# GiottoClass 0.2.2
+
+## bug fixes
+- fix `createGiottoPolygonsFromMask()` IDs being applied out of sync to mask values
+- remove unused `fix_multipart` param in `createGiottoPolygonsFromMask()`
+
+## enhancements
+- `createGiottoPolygonsFromMask()` now has `ID_fmt` param for finer control of automatic poly_ID generation
+
+
 # GiottoClass 0.2.1 (2024/02/28)
 
 ## breaking changes